+++ inventory_tools/hooks.py
@@ -51,7 +51,7 @@ doctype_js = {
 	"Work Order": "public/js/custom/work_order_custom.js",
 	"Workstation": "public/js/custom/workstation_custom.js",
 }
-# doctype_list_js = {"doctype" : "public/js/doctype_list.js"}
+doctype_list_js = {"Work Order": "public/js/custom/work_order_list.js"}
 doctype_tree_js = {"Warehouse": "public/js/custom/warehouse_tree.js"}
 # doctype_calendar_js = {"doctype" : "public/js/doctype_calendar.js"}
